wip remove keystore
