+++ api/common_args_responses.go
@@ -22,12 +22,6 @@ type JSONTxID struct {
 	TxID ids.ID `json:"txID"`
 }
 
-// UserPass contains a username and a password
-type UserPass struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
 // JSONAddress contains an address
 type JSONAddress struct {
 	Address string `json:"address"`
@@ -52,16 +46,6 @@ type JSONTxIDChangeAddr struct {
 // JSONFromAddrs is a list of addresses to send funds from
 type JSONFromAddrs struct {
 	From []string `json:"from"`
-}
-
-// JSONSpendHeader is 3 arguments to a method that spends (including those with tx fees)
-// 1) The username/password
-// 2) The addresses used in the method
-// 3) The address to send change to
-type JSONSpendHeader struct {
-	UserPass
-	JSONFromAddrs
-	JSONChangeAddr
 }
 
 // GetBlockArgs is the parameters supplied to the GetBlock API
