--- api/keystore/codec.go
@@ -1,26 +1,0 @@
-// Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package keystore
-
-import (
-	"github.com/ava-labs/avalanchego/codec"
-	"github.com/ava-labs/avalanchego/codec/linearcodec"
-	"github.com/ava-labs/avalanchego/utils/units"
-)
-
-const (
-	CodecVersion = 0
-
-	maxPackerSize = 1 * units.GiB // max size, in bytes, of something being marshalled by Marshal()
-)
-
-var Codec codec.Manager
-
-func init() {
-	lc := linearcodec.NewDefault()
-	Codec = codec.NewManager(maxPackerSize)
-	if err := Codec.RegisterCodec(CodecVersion, lc); err != nil {
-		panic(err)
-	}
-}
