+++ config/keys.go
@@ -145,7 +145,6 @@ const (
 	TrackSubnetsKey                                    = "track-subnets"
 	AdminAPIEnabledKey                                 = "api-admin-enabled"
 	InfoAPIEnabledKey                                  = "api-info-enabled"
-	KeystoreAPIEnabledKey                              = "api-keystore-enabled"
 	MetricsAPIEnabledKey                               = "api-metrics-enabled"
 	HealthAPIEnabledKey                                = "api-health-enabled"
 	MeterVMsEnabledKey                                 = "meter-vms-enabled"
