--- proto/keystore/keystore.proto
@@ -1,23 +1,0 @@
-syntax = "proto3";
-
-package keystore;
-
-option go_package = "github.com/ava-labs/avalanchego/proto/pb/keystore";
-
-service Keystore {
-  rpc GetDatabase(GetDatabaseRequest) returns (GetDatabaseResponse);
-}
-
-message GetDatabaseRequest {
-  string username = 1;
-  string password = 2;
-}
-
-message GetDatabaseResponse {
-  // reserved for backward compatibility
-  // avalanchego <=v1.7.9 used the field "1" as an id to identify the gRPC server
-  // address which served the Database service via the now removed service broker
-  reserved 1;
-  // server_addr is the address of the gRPC server hosting the Database service
-  string server_addr = 2;
-}
