+++ proto/vm/vm.proto
@@ -108,9 +108,8 @@ message InitializeRequest {
   bytes upgrade_bytes = 11;
   bytes config_bytes = 12;
   string db_server_addr = 13;
-  // server_addr is the address of the gRPC server which serves
-  // the messenger, keystore, shared memory, blockchain alias,
-  // subnet alias, and appSender services
+  // server_addr is the address of the gRPC server which serves the messenger,
+  // shared memory, blockchain alias, subnet alias, and appSender services
   string server_addr = 14;
   // network_upgrades_bytes is the json encoded network upgrades
   NetworkUpgrades network_upgrades = 15;
