+++ vms/avm/client.go
@@ -13,7 +13,6 @@ import (
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow/choices"
 	"github.com/ava-labs/avalanchego/utils/constants"
-	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/avalanchego/utils/formatting"
 	"github.com/ava-labs/avalanchego/utils/formatting/address"
 	"github.com/ava-labs/avalanchego/utils/json"
@@ -76,147 +75,6 @@ type Client interface {
 
 	// GetTxFee returns the cost to issue certain transactions
 	GetTxFee(context.Context, ...rpc.Option) (uint64, uint64, error)
-
-	// CreateAsset creates a new asset and returns its assetID
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	CreateAsset(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		name string,
-		symbol string,
-		denomination byte,
-		holders []*ClientHolder,
-		minters []ClientOwners,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// CreateFixedCapAsset creates a new fixed cap asset and returns its assetID
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	CreateFixedCapAsset(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		name string,
-		symbol string,
-		denomination byte,
-		holders []*ClientHolder,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// CreateVariableCapAsset creates a new variable cap asset and returns its assetID
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	CreateVariableCapAsset(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		name string,
-		symbol string,
-		denomination byte,
-		minters []ClientOwners,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// CreateNFTAsset creates a new NFT asset and returns its assetID
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	CreateNFTAsset(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		name string,
-		symbol string,
-		minters []ClientOwners,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// CreateAddress creates a new address controlled by [user]
-	//
-	// Deprecated: Keys should no longer be stored on the node.
-	CreateAddress(ctx context.Context, user api.UserPass, options ...rpc.Option) (ids.ShortID, error)
-	// ListAddresses returns all addresses on this chain controlled by [user]
-	//
-	// Deprecated: Keys should no longer be stored on the node.
-	ListAddresses(ctx context.Context, user api.UserPass, options ...rpc.Option) ([]ids.ShortID, error)
-	// ExportKey returns the private key corresponding to [addr] controlled by [user]
-	//
-	// Deprecated: Keys should no longer be stored on the node.
-	ExportKey(ctx context.Context, user api.UserPass, addr ids.ShortID, options ...rpc.Option) (*secp256k1.PrivateKey, error)
-	// ImportKey imports [privateKey] to [user]
-	//
-	// Deprecated: Keys should no longer be stored on the node.
-	ImportKey(ctx context.Context, user api.UserPass, privateKey *secp256k1.PrivateKey, options ...rpc.Option) (ids.ShortID, error)
-	// Mint [amount] of [assetID] to be owned by [to]
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	Mint(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		amount uint64,
-		assetID string,
-		to ids.ShortID,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// SendNFT sends an NFT and returns the ID of the newly created transaction
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	SendNFT(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		assetID string,
-		groupID uint32,
-		to ids.ShortID,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// MintNFT issues a MintNFT transaction and returns the ID of the newly created transaction
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	MintNFT(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		assetID string,
-		payload []byte,
-		to ids.ShortID,
-		options ...rpc.Option,
-	) (ids.ID, error)
-	// Import sends an import transaction to import funds from [sourceChain] and
-	// returns the ID of the newly created transaction
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	Import(ctx context.Context, user api.UserPass, to ids.ShortID, sourceChain string, options ...rpc.Option) (ids.ID, error) // Export sends an asset from this chain to the P/C-Chain.
-	// After this tx is accepted, the AVAX must be imported to the P/C-chain with an importTx.
-	// Returns the ID of the newly created atomic transaction
-	//
-	// Deprecated: Transactions should be issued using the
-	// `avalanchego/wallet/chain/x.Wallet` utility.
-	Export(
-		ctx context.Context,
-		user api.UserPass,
-		from []ids.ShortID,
-		changeAddr ids.ShortID,
-		amount uint64,
-		to ids.ShortID,
-		toChainIDAlias string,
-		assetID string,
-		options ...rpc.Option,
-	) (ids.ID, error)
 }
 
 // implementation for an AVM client for interacting with avm [chain]
@@ -393,367 +251,6 @@ func (c *client) GetTxFee(ctx context.Context, options ...rpc.Option) (uint64, u
 	res := &GetTxFeeReply{}
 	err := c.requester.SendRequest(ctx, "avm.getTxFee", struct{}{}, res, options...)
 	return uint64(res.TxFee), uint64(res.CreateAssetTxFee), err
-}
-
-// ClientHolder describes how much an address owns of an asset
-type ClientHolder struct {
-	Amount  uint64
-	Address ids.ShortID
-}
-
-// ClientOwners describes who can perform an action
-type ClientOwners struct {
-	Threshold uint32
-	Minters   []ids.ShortID
-}
-
-func (c *client) CreateAsset(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	name string,
-	symbol string,
-	denomination byte,
-	clientHolders []*ClientHolder,
-	clientMinters []ClientOwners,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &FormattedAssetID{}
-	holders := make([]*Holder, len(clientHolders))
-	for i, clientHolder := range clientHolders {
-		holders[i] = &Holder{
-			Amount:  json.Uint64(clientHolder.Amount),
-			Address: clientHolder.Address.String(),
-		}
-	}
-	minters := make([]Owners, len(clientMinters))
-	for i, clientMinter := range clientMinters {
-		minters[i] = Owners{
-			Threshold: json.Uint32(clientMinter.Threshold),
-			Minters:   ids.ShortIDsToStrings(clientMinter.Minters),
-		}
-	}
-	err := c.requester.SendRequest(ctx, "avm.createAsset", &CreateAssetArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Name:           name,
-		Symbol:         symbol,
-		Denomination:   denomination,
-		InitialHolders: holders,
-		MinterSets:     minters,
-	}, res, options...)
-	return res.AssetID, err
-}
-
-func (c *client) CreateFixedCapAsset(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	name string,
-	symbol string,
-	denomination byte,
-	clientHolders []*ClientHolder,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &FormattedAssetID{}
-	holders := make([]*Holder, len(clientHolders))
-	for i, clientHolder := range clientHolders {
-		holders[i] = &Holder{
-			Amount:  json.Uint64(clientHolder.Amount),
-			Address: clientHolder.Address.String(),
-		}
-	}
-	err := c.requester.SendRequest(ctx, "avm.createAsset", &CreateAssetArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Name:           name,
-		Symbol:         symbol,
-		Denomination:   denomination,
-		InitialHolders: holders,
-	}, res, options...)
-	return res.AssetID, err
-}
-
-func (c *client) CreateVariableCapAsset(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	name string,
-	symbol string,
-	denomination byte,
-	clientMinters []ClientOwners,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &FormattedAssetID{}
-	minters := make([]Owners, len(clientMinters))
-	for i, clientMinter := range clientMinters {
-		minters[i] = Owners{
-			Threshold: json.Uint32(clientMinter.Threshold),
-			Minters:   ids.ShortIDsToStrings(clientMinter.Minters),
-		}
-	}
-	err := c.requester.SendRequest(ctx, "avm.createAsset", &CreateAssetArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Name:         name,
-		Symbol:       symbol,
-		Denomination: denomination,
-		MinterSets:   minters,
-	}, res, options...)
-	return res.AssetID, err
-}
-
-func (c *client) CreateNFTAsset(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	name string,
-	symbol string,
-	clientMinters []ClientOwners,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &FormattedAssetID{}
-	minters := make([]Owners, len(clientMinters))
-	for i, clientMinter := range clientMinters {
-		minters[i] = Owners{
-			Threshold: json.Uint32(clientMinter.Threshold),
-			Minters:   ids.ShortIDsToStrings(clientMinter.Minters),
-		}
-	}
-	err := c.requester.SendRequest(ctx, "avm.createNFTAsset", &CreateNFTAssetArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Name:       name,
-		Symbol:     symbol,
-		MinterSets: minters,
-	}, res, options...)
-	return res.AssetID, err
-}
-
-func (c *client) CreateAddress(ctx context.Context, user api.UserPass, options ...rpc.Option) (ids.ShortID, error) {
-	res := &api.JSONAddress{}
-	err := c.requester.SendRequest(ctx, "avm.createAddress", &user, res, options...)
-	if err != nil {
-		return ids.ShortID{}, err
-	}
-	return address.ParseToID(res.Address)
-}
-
-func (c *client) ListAddresses(ctx context.Context, user api.UserPass, options ...rpc.Option) ([]ids.ShortID, error) {
-	res := &api.JSONAddresses{}
-	err := c.requester.SendRequest(ctx, "avm.listAddresses", &user, res, options...)
-	if err != nil {
-		return nil, err
-	}
-	return address.ParseToIDs(res.Addresses)
-}
-
-func (c *client) ExportKey(ctx context.Context, user api.UserPass, addr ids.ShortID, options ...rpc.Option) (*secp256k1.PrivateKey, error) {
-	res := &ExportKeyReply{}
-	err := c.requester.SendRequest(ctx, "avm.exportKey", &ExportKeyArgs{
-		UserPass: user,
-		Address:  addr.String(),
-	}, res, options...)
-	return res.PrivateKey, err
-}
-
-func (c *client) ImportKey(ctx context.Context, user api.UserPass, privateKey *secp256k1.PrivateKey, options ...rpc.Option) (ids.ShortID, error) {
-	res := &api.JSONAddress{}
-	err := c.requester.SendRequest(ctx, "avm.importKey", &ImportKeyArgs{
-		UserPass:   user,
-		PrivateKey: privateKey,
-	}, res, options...)
-	if err != nil {
-		return ids.ShortID{}, err
-	}
-	return address.ParseToID(res.Address)
-}
-
-func (c *client) Send(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	amount uint64,
-	assetID string,
-	to ids.ShortID,
-	memo string,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	err := c.requester.SendRequest(ctx, "avm.send", &SendArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		SendOutput: SendOutput{
-			Amount:  json.Uint64(amount),
-			AssetID: assetID,
-			To:      to.String(),
-		},
-		Memo: memo,
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) SendMultiple(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	clientOutputs []ClientSendOutput,
-	memo string,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	outputs := make([]SendOutput, len(clientOutputs))
-	for i, clientOutput := range clientOutputs {
-		outputs[i] = SendOutput{
-			Amount:  json.Uint64(clientOutput.Amount),
-			AssetID: clientOutput.AssetID,
-			To:      clientOutput.To.String(),
-		}
-	}
-	err := c.requester.SendRequest(ctx, "avm.sendMultiple", &SendMultipleArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Outputs: outputs,
-		Memo:    memo,
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) Mint(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	amount uint64,
-	assetID string,
-	to ids.ShortID,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	err := c.requester.SendRequest(ctx, "avm.mint", &MintArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Amount:  json.Uint64(amount),
-		AssetID: assetID,
-		To:      to.String(),
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) SendNFT(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	assetID string,
-	groupID uint32,
-	to ids.ShortID,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	err := c.requester.SendRequest(ctx, "avm.sendNFT", &SendNFTArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		AssetID: assetID,
-		GroupID: json.Uint32(groupID),
-		To:      to.String(),
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) MintNFT(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	assetID string,
-	payload []byte,
-	to ids.ShortID,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	payloadStr, err := formatting.Encode(formatting.Hex, payload)
-	if err != nil {
-		return ids.Empty, err
-	}
-	res := &api.JSONTxID{}
-	err = c.requester.SendRequest(ctx, "avm.mintNFT", &MintNFTArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		AssetID:  assetID,
-		Payload:  payloadStr,
-		To:       to.String(),
-		Encoding: formatting.Hex,
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) Import(ctx context.Context, user api.UserPass, to ids.ShortID, sourceChain string, options ...rpc.Option) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	err := c.requester.SendRequest(ctx, "avm.import", &ImportArgs{
-		UserPass:    user,
-		To:          to.String(),
-		SourceChain: sourceChain,
-	}, res, options...)
-	return res.TxID, err
-}
-
-func (c *client) Export(
-	ctx context.Context,
-	user api.UserPass,
-	from []ids.ShortID,
-	changeAddr ids.ShortID,
-	amount uint64,
-	to ids.ShortID,
-	targetChain string,
-	assetID string,
-	options ...rpc.Option,
-) (ids.ID, error) {
-	res := &api.JSONTxID{}
-	err := c.requester.SendRequest(ctx, "avm.export", &ExportArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass:       user,
-			JSONFromAddrs:  api.JSONFromAddrs{From: ids.ShortIDsToStrings(from)},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddr.String()},
-		},
-		Amount:      json.Uint64(amount),
-		TargetChain: targetChain,
-		To:          to.String(),
-		AssetID:     assetID,
-	}, res, options...)
-	return res.TxID, err
 }
 
 func AwaitTxAccepted(
