+++ vms/avm/service.go
@@ -16,16 +16,11 @@ import (
 	"github.com/ava-labs/avalanchego/database"
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow/choices"
-	"github.com/ava-labs/avalanchego/utils"
-	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/avalanchego/utils/formatting"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/vms/avm/txs"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
-	"github.com/ava-labs/avalanchego/vms/components/keystore"
-	"github.com/ava-labs/avalanchego/vms/components/verify"
-	"github.com/ava-labs/avalanchego/vms/nftfx"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 
 	avajson "github.com/ava-labs/avalanchego/utils/json"
@@ -686,1348 +681,4 @@ func (s *Service) GetTxFee(_ *http.Request, _ *struct{}, reply *GetTxFeeReply) e
 	reply.TxFee = avajson.Uint64(s.vm.TxFee)
 	reply.CreateAssetTxFee = avajson.Uint64(s.vm.CreateAssetTxFee)
 	return nil
-}
-
-// Holder describes how much an address owns of an asset
-type Holder struct {
-	Amount  avajson.Uint64 `json:"amount"`
-	Address string         `json:"address"`
-}
-
-// Owners describes who can perform an action
-type Owners struct {
-	Threshold avajson.Uint32 `json:"threshold"`
-	Minters   []string       `json:"minters"`
-}
-
-// CreateAssetArgs are arguments for passing into CreateAsset
-type CreateAssetArgs struct {
-	api.JSONSpendHeader           // User, password, from addrs, change addr
-	Name                string    `json:"name"`
-	Symbol              string    `json:"symbol"`
-	Denomination        byte      `json:"denomination"`
-	InitialHolders      []*Holder `json:"initialHolders"`
-	MinterSets          []Owners  `json:"minterSets"`
-}
-
-// AssetIDChangeAddr is an asset ID and a change address
-type AssetIDChangeAddr struct {
-	FormattedAssetID
-	api.JSONChangeAddr
-}
-
-// CreateAsset returns ID of the newly created asset
-func (s *Service) CreateAsset(_ *http.Request, args *CreateAssetArgs, reply *AssetIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "createAsset"),
-		logging.UserString("name", args.Name),
-		logging.UserString("symbol", args.Symbol),
-		zap.Int("numInitialHolders", len(args.InitialHolders)),
-		zap.Int("numMinters", len(args.MinterSets)),
-	)
-
-	tx, changeAddr, err := s.buildCreateAssetTx(args)
-	if err != nil {
-		return err
-	}
-
-	assetID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.AssetID = assetID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildCreateAssetTx(args *CreateAssetArgs) (*txs.Tx, ids.ShortID, error) {
-	if len(args.InitialHolders) == 0 && len(args.MinterSets) == 0 {
-		return nil, ids.ShortEmpty, errNoHoldersOrMinters
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(kc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(kc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amountsSpent, ins, keys, err := s.vm.Spend(
-		utxos,
-		kc,
-		map[ids.ID]uint64{
-			s.vm.feeAssetID: s.vm.CreateAssetTxFee,
-		},
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	outs := []*avax.TransferableOutput{}
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent > s.vm.CreateAssetTxFee {
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: s.vm.feeAssetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: amountSpent - s.vm.CreateAssetTxFee,
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{changeAddr},
-				},
-			},
-		})
-	}
-
-	initialState := &txs.InitialState{
-		FxIndex: 0, // TODO: Should lookup secp256k1fx FxID
-		Outs:    make([]verify.State, 0, len(args.InitialHolders)+len(args.MinterSets)),
-	}
-	for _, holder := range args.InitialHolders {
-		addr, err := avax.ParseServiceAddress(s.vm, holder.Address)
-		if err != nil {
-			return nil, ids.ShortEmpty, err
-		}
-		initialState.Outs = append(initialState.Outs, &secp256k1fx.TransferOutput{
-			Amt: uint64(holder.Amount),
-			OutputOwners: secp256k1fx.OutputOwners{
-				Threshold: 1,
-				Addrs:     []ids.ShortID{addr},
-			},
-		})
-	}
-	for _, owner := range args.MinterSets {
-		minter := &secp256k1fx.MintOutput{
-			OutputOwners: secp256k1fx.OutputOwners{
-				Threshold: uint32(owner.Threshold),
-				Addrs:     make([]ids.ShortID, 0, len(owner.Minters)),
-			},
-		}
-		minterAddrsSet, err := avax.ParseServiceAddresses(s.vm, owner.Minters)
-		if err != nil {
-			return nil, ids.ShortEmpty, err
-		}
-		minter.Addrs = minterAddrsSet.List()
-		utils.Sort(minter.Addrs)
-		initialState.Outs = append(initialState.Outs, minter)
-	}
-
-	codec := s.vm.parser.Codec()
-	initialState.Sort(codec)
-
-	tx := &txs.Tx{Unsigned: &txs.CreateAssetTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		Name:         args.Name,
-		Symbol:       args.Symbol,
-		Denomination: args.Denomination,
-		States:       []*txs.InitialState{initialState},
-	}}
-	return tx, changeAddr, tx.SignSECP256K1Fx(codec, keys)
-}
-
-// CreateFixedCapAsset returns ID of the newly created asset
-func (s *Service) CreateFixedCapAsset(r *http.Request, args *CreateAssetArgs, reply *AssetIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "createFixedCapAsset"),
-		logging.UserString("name", args.Name),
-		logging.UserString("symbol", args.Symbol),
-		zap.Int("numInitialHolders", len(args.InitialHolders)),
-	)
-
-	return s.CreateAsset(r, args, reply)
-}
-
-// CreateVariableCapAsset returns ID of the newly created asset
-func (s *Service) CreateVariableCapAsset(r *http.Request, args *CreateAssetArgs, reply *AssetIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "createVariableCapAsset"),
-		logging.UserString("name", args.Name),
-		logging.UserString("symbol", args.Symbol),
-		zap.Int("numMinters", len(args.MinterSets)),
-	)
-
-	return s.CreateAsset(r, args, reply)
-}
-
-// CreateNFTAssetArgs are arguments for passing into CreateNFTAsset requests
-type CreateNFTAssetArgs struct {
-	api.JSONSpendHeader          // User, password, from addrs, change addr
-	Name                string   `json:"name"`
-	Symbol              string   `json:"symbol"`
-	MinterSets          []Owners `json:"minterSets"`
-}
-
-// CreateNFTAsset returns ID of the newly created asset
-func (s *Service) CreateNFTAsset(_ *http.Request, args *CreateNFTAssetArgs, reply *AssetIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "createNFTAsset"),
-		logging.UserString("name", args.Name),
-		logging.UserString("symbol", args.Symbol),
-		zap.Int("numMinters", len(args.MinterSets)),
-	)
-
-	tx, changeAddr, err := s.buildCreateNFTAsset(args)
-	if err != nil {
-		return err
-	}
-
-	assetID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.AssetID = assetID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildCreateNFTAsset(args *CreateNFTAssetArgs) (*txs.Tx, ids.ShortID, error) {
-	if len(args.MinterSets) == 0 {
-		return nil, ids.ShortEmpty, errNoMinters
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(kc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(kc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amountsSpent, ins, keys, err := s.vm.Spend(
-		utxos,
-		kc,
-		map[ids.ID]uint64{
-			s.vm.feeAssetID: s.vm.CreateAssetTxFee,
-		},
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	outs := []*avax.TransferableOutput{}
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent > s.vm.CreateAssetTxFee {
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: s.vm.feeAssetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: amountSpent - s.vm.CreateAssetTxFee,
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{changeAddr},
-				},
-			},
-		})
-	}
-
-	initialState := &txs.InitialState{
-		FxIndex: 1, // TODO: Should lookup nftfx FxID
-		Outs:    make([]verify.State, 0, len(args.MinterSets)),
-	}
-	for i, owner := range args.MinterSets {
-		minter := &nftfx.MintOutput{
-			GroupID: uint32(i),
-			OutputOwners: secp256k1fx.OutputOwners{
-				Threshold: uint32(owner.Threshold),
-			},
-		}
-		minterAddrsSet, err := avax.ParseServiceAddresses(s.vm, owner.Minters)
-		if err != nil {
-			return nil, ids.ShortEmpty, err
-		}
-		minter.Addrs = minterAddrsSet.List()
-		utils.Sort(minter.Addrs)
-		initialState.Outs = append(initialState.Outs, minter)
-	}
-
-	codec := s.vm.parser.Codec()
-	initialState.Sort(codec)
-
-	tx := &txs.Tx{Unsigned: &txs.CreateAssetTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		Name:         args.Name,
-		Symbol:       args.Symbol,
-		Denomination: 0, // NFTs are non-fungible
-		States:       []*txs.InitialState{initialState},
-	}}
-	return tx, changeAddr, tx.SignSECP256K1Fx(codec, keys)
-}
-
-// CreateAddress creates an address for the user [args.Username]
-func (s *Service) CreateAddress(_ *http.Request, args *api.UserPass, reply *api.JSONAddress) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "createAddress"),
-		logging.UserString("username", args.Username),
-	)
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	user, err := keystore.NewUserFromKeystore(s.vm.ctx.Keystore, args.Username, args.Password)
-	if err != nil {
-		return err
-	}
-	defer user.Close()
-
-	sk, err := keystore.NewKey(user)
-	if err != nil {
-		return err
-	}
-
-	reply.Address, err = s.vm.FormatLocalAddress(sk.PublicKey().Address())
-	if err != nil {
-		return fmt.Errorf("problem formatting address: %w", err)
-	}
-
-	// Return an error if the DB can't close, this will execute before the above
-	// db close.
-	return user.Close()
-}
-
-// ListAddresses returns all of the addresses controlled by user [args.Username]
-func (s *Service) ListAddresses(_ *http.Request, args *api.UserPass, response *api.JSONAddresses) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "listAddresses"),
-		logging.UserString("username", args.Username),
-	)
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	user, err := keystore.NewUserFromKeystore(s.vm.ctx.Keystore, args.Username, args.Password)
-	if err != nil {
-		return err
-	}
-
-	response.Addresses = []string{}
-
-	addresses, err := user.GetAddresses()
-	if err != nil {
-		// An error fetching the addresses may just mean that the user has no
-		// addresses.
-		return user.Close()
-	}
-
-	for _, address := range addresses {
-		addr, err := s.vm.FormatLocalAddress(address)
-		if err != nil {
-			// Drop any potential error closing the database to report the
-			// original error
-			_ = user.Close()
-			return fmt.Errorf("problem formatting address: %w", err)
-		}
-		response.Addresses = append(response.Addresses, addr)
-	}
-	return user.Close()
-}
-
-// ExportKeyArgs are arguments for ExportKey
-type ExportKeyArgs struct {
-	api.UserPass
-	Address string `json:"address"`
-}
-
-// ExportKeyReply is the response for ExportKey
-type ExportKeyReply struct {
-	// The decrypted PrivateKey for the Address provided in the arguments
-	PrivateKey *secp256k1.PrivateKey `json:"privateKey"`
-}
-
-// ExportKey returns a private key from the provided user
-func (s *Service) ExportKey(_ *http.Request, args *ExportKeyArgs, reply *ExportKeyReply) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "exportKey"),
-		logging.UserString("username", args.Username),
-	)
-
-	addr, err := avax.ParseServiceAddress(s.vm, args.Address)
-	if err != nil {
-		return fmt.Errorf("problem parsing address %q: %w", args.Address, err)
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	user, err := keystore.NewUserFromKeystore(s.vm.ctx.Keystore, args.Username, args.Password)
-	if err != nil {
-		return err
-	}
-
-	reply.PrivateKey, err = user.GetKey(addr)
-	if err != nil {
-		// Drop any potential error closing the database to report the original
-		// error
-		_ = user.Close()
-		return fmt.Errorf("problem retrieving private key: %w", err)
-	}
-	return user.Close()
-}
-
-// ImportKeyArgs are arguments for ImportKey
-type ImportKeyArgs struct {
-	api.UserPass
-	PrivateKey *secp256k1.PrivateKey `json:"privateKey"`
-}
-
-// ImportKeyReply is the response for ImportKey
-type ImportKeyReply struct {
-	// The address controlled by the PrivateKey provided in the arguments
-	Address string `json:"address"`
-}
-
-// ImportKey adds a private key to the provided user
-func (s *Service) ImportKey(_ *http.Request, args *ImportKeyArgs, reply *api.JSONAddress) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "importKey"),
-		logging.UserString("username", args.Username),
-	)
-
-	if args.PrivateKey == nil {
-		return errMissingPrivateKey
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	user, err := keystore.NewUserFromKeystore(s.vm.ctx.Keystore, args.Username, args.Password)
-	if err != nil {
-		return err
-	}
-	defer user.Close()
-
-	if err := user.PutKeys(args.PrivateKey); err != nil {
-		return fmt.Errorf("problem saving key %w", err)
-	}
-
-	newAddress := args.PrivateKey.PublicKey().Address()
-	reply.Address, err = s.vm.FormatLocalAddress(newAddress)
-	if err != nil {
-		return fmt.Errorf("problem formatting address: %w", err)
-	}
-
-	return user.Close()
-}
-
-// SendOutput specifies that [Amount] of asset [AssetID] be sent to [To]
-type SendOutput struct {
-	// The amount of funds to send
-	Amount avajson.Uint64 `json:"amount"`
-
-	// ID of the asset being sent
-	AssetID string `json:"assetID"`
-
-	// Address of the recipient
-	To string `json:"to"`
-}
-
-// SendArgs are arguments for passing into Send requests
-type SendArgs struct {
-	// User, password, from addrs, change addr
-	api.JSONSpendHeader
-
-	// The amount, assetID, and destination to send funds to
-	SendOutput
-
-	// Memo field
-	Memo string `json:"memo"`
-}
-
-// SendMultipleArgs are arguments for passing into SendMultiple requests
-type SendMultipleArgs struct {
-	// User, password, from addrs, change addr
-	api.JSONSpendHeader
-
-	// The outputs of the transaction
-	Outputs []SendOutput `json:"outputs"`
-
-	// Memo field
-	Memo string `json:"memo"`
-}
-
-// Send returns the ID of the newly created transaction
-func (s *Service) Send(r *http.Request, args *SendArgs, reply *api.JSONTxIDChangeAddr) error {
-	return s.SendMultiple(r, &SendMultipleArgs{
-		JSONSpendHeader: args.JSONSpendHeader,
-		Outputs:         []SendOutput{args.SendOutput},
-		Memo:            args.Memo,
-	}, reply)
-}
-
-// SendMultiple sends a transaction with multiple outputs.
-func (s *Service) SendMultiple(_ *http.Request, args *SendMultipleArgs, reply *api.JSONTxIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "sendMultiple"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, changeAddr, err := s.buildSendMultiple(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildSendMultiple(args *SendMultipleArgs) (*txs.Tx, ids.ShortID, error) {
-	// Validate the memo field
-	memoBytes := []byte(args.Memo)
-	if l := len(memoBytes); l > avax.MaxMemoSize {
-		return nil, ids.ShortEmpty, fmt.Errorf("max memo length is %d but provided memo field is length %d", avax.MaxMemoSize, l)
-	} else if len(args.Outputs) == 0 {
-		return nil, ids.ShortEmpty, errNoOutputs
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Load user's UTXOs/keys
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(kc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(kc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Calculate required input amounts and create the desired outputs
-	// String repr. of asset ID --> asset ID
-	assetIDs := make(map[string]ids.ID)
-	// Asset ID --> amount of that asset being sent
-	amounts := make(map[ids.ID]uint64)
-	// Outputs of our tx
-	outs := []*avax.TransferableOutput{}
-	for _, output := range args.Outputs {
-		if output.Amount == 0 {
-			return nil, ids.ShortEmpty, errZeroAmount
-		}
-		assetID, ok := assetIDs[output.AssetID] // Asset ID of next output
-		if !ok {
-			assetID, err = s.vm.lookupAssetID(output.AssetID)
-			if err != nil {
-				return nil, ids.ShortEmpty, fmt.Errorf("couldn't find asset %s", output.AssetID)
-			}
-			assetIDs[output.AssetID] = assetID
-		}
-		currentAmount := amounts[assetID]
-		newAmount, err := safemath.Add(currentAmount, uint64(output.Amount))
-		if err != nil {
-			return nil, ids.ShortEmpty, fmt.Errorf("problem calculating required spend amount: %w", err)
-		}
-		amounts[assetID] = newAmount
-
-		// Parse the to address
-		to, err := avax.ParseServiceAddress(s.vm, output.To)
-		if err != nil {
-			return nil, ids.ShortEmpty, fmt.Errorf("problem parsing to address %q: %w", output.To, err)
-		}
-
-		// Create the Output
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: assetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: uint64(output.Amount),
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{to},
-				},
-			},
-		})
-	}
-
-	amountsWithFee := make(map[ids.ID]uint64, len(amounts)+1)
-	for assetID, amount := range amounts {
-		amountsWithFee[assetID] = amount
-	}
-
-	amountWithFee, err := safemath.Add(amounts[s.vm.feeAssetID], s.vm.TxFee)
-	if err != nil {
-		return nil, ids.ShortEmpty, fmt.Errorf("problem calculating required spend amount: %w", err)
-	}
-	amountsWithFee[s.vm.feeAssetID] = amountWithFee
-
-	amountsSpent, ins, keys, err := s.vm.Spend(
-		utxos,
-		kc,
-		amountsWithFee,
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Add the required change outputs
-	for assetID, amountWithFee := range amountsWithFee {
-		amountSpent := amountsSpent[assetID]
-
-		if amountSpent > amountWithFee {
-			outs = append(outs, &avax.TransferableOutput{
-				Asset: avax.Asset{ID: assetID},
-				Out: &secp256k1fx.TransferOutput{
-					Amt: amountSpent - amountWithFee,
-					OutputOwners: secp256k1fx.OutputOwners{
-						Locktime:  0,
-						Threshold: 1,
-						Addrs:     []ids.ShortID{changeAddr},
-					},
-				},
-			})
-		}
-	}
-
-	codec := s.vm.parser.Codec()
-	avax.SortTransferableOutputs(outs, codec)
-
-	tx := &txs.Tx{Unsigned: &txs.BaseTx{BaseTx: avax.BaseTx{
-		NetworkID:    s.vm.ctx.NetworkID,
-		BlockchainID: s.vm.ctx.ChainID,
-		Outs:         outs,
-		Ins:          ins,
-		Memo:         memoBytes,
-	}}}
-	return tx, changeAddr, tx.SignSECP256K1Fx(codec, keys)
-}
-
-// MintArgs are arguments for passing into Mint requests
-type MintArgs struct {
-	api.JSONSpendHeader                // User, password, from addrs, change addr
-	Amount              avajson.Uint64 `json:"amount"`
-	AssetID             string         `json:"assetID"`
-	To                  string         `json:"to"`
-}
-
-// Mint issues a transaction that mints more of the asset
-func (s *Service) Mint(_ *http.Request, args *MintArgs, reply *api.JSONTxIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "mint"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, changeAddr, err := s.buildMint(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildMint(args *MintArgs) (*txs.Tx, ids.ShortID, error) {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "mint"),
-		logging.UserString("username", args.Username),
-	)
-
-	if args.Amount == 0 {
-		return nil, ids.ShortEmpty, errInvalidMintAmount
-	}
-
-	assetID, err := s.vm.lookupAssetID(args.AssetID)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	to, err := avax.ParseServiceAddress(s.vm, args.To)
-	if err != nil {
-		return nil, ids.ShortEmpty, fmt.Errorf("problem parsing to address %q: %w", args.To, err)
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	feeUTXOs, feeKc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(feeKc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(feeKc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amountsSpent, ins, keys, err := s.vm.Spend(
-		feeUTXOs,
-		feeKc,
-		map[ids.ID]uint64{
-			s.vm.feeAssetID: s.vm.TxFee,
-		},
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	outs := []*avax.TransferableOutput{}
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent > s.vm.TxFee {
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: s.vm.feeAssetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: amountSpent - s.vm.TxFee,
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{changeAddr},
-				},
-			},
-		})
-	}
-
-	// Get all UTXOs/keys for the user
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, nil)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	ops, opKeys, err := s.vm.Mint(
-		utxos,
-		kc,
-		map[ids.ID]uint64{
-			assetID: uint64(args.Amount),
-		},
-		to,
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-	keys = append(keys, opKeys...)
-
-	tx := &txs.Tx{Unsigned: &txs.OperationTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		Ops: ops,
-	}}
-	return tx, changeAddr, tx.SignSECP256K1Fx(s.vm.parser.Codec(), keys)
-}
-
-// SendNFTArgs are arguments for passing into SendNFT requests
-type SendNFTArgs struct {
-	api.JSONSpendHeader                // User, password, from addrs, change addr
-	AssetID             string         `json:"assetID"`
-	GroupID             avajson.Uint32 `json:"groupID"`
-	To                  string         `json:"to"`
-}
-
-// SendNFT sends an NFT
-func (s *Service) SendNFT(_ *http.Request, args *SendNFTArgs, reply *api.JSONTxIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "sendNFT"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, changeAddr, err := s.buildSendNFT(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildSendNFT(args *SendNFTArgs) (*txs.Tx, ids.ShortID, error) {
-	// Parse the asset ID
-	assetID, err := s.vm.lookupAssetID(args.AssetID)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the to address
-	to, err := avax.ParseServiceAddress(s.vm, args.To)
-	if err != nil {
-		return nil, ids.ShortEmpty, fmt.Errorf("problem parsing to address %q: %w", args.To, err)
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(kc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(kc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amountsSpent, ins, secpKeys, err := s.vm.Spend(
-		utxos,
-		kc,
-		map[ids.ID]uint64{
-			s.vm.feeAssetID: s.vm.TxFee,
-		},
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	outs := []*avax.TransferableOutput{}
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent > s.vm.TxFee {
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: s.vm.feeAssetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: amountSpent - s.vm.TxFee,
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{changeAddr},
-				},
-			},
-		})
-	}
-
-	ops, nftKeys, err := s.vm.SpendNFT(
-		utxos,
-		kc,
-		assetID,
-		uint32(args.GroupID),
-		to,
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	tx := &txs.Tx{Unsigned: &txs.OperationTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		Ops: ops,
-	}}
-
-	codec := s.vm.parser.Codec()
-	if err := tx.SignSECP256K1Fx(codec, secpKeys); err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-	return tx, changeAddr, tx.SignNFTFx(codec, nftKeys)
-}
-
-// MintNFTArgs are arguments for passing into MintNFT requests
-type MintNFTArgs struct {
-	api.JSONSpendHeader                     // User, password, from addrs, change addr
-	AssetID             string              `json:"assetID"`
-	Payload             string              `json:"payload"`
-	To                  string              `json:"to"`
-	Encoding            formatting.Encoding `json:"encoding"`
-}
-
-// MintNFT issues a MintNFT transaction and returns the ID of the newly created transaction
-func (s *Service) MintNFT(_ *http.Request, args *MintNFTArgs, reply *api.JSONTxIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "mintNFT"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, changeAddr, err := s.buildMintNFT(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildMintNFT(args *MintNFTArgs) (*txs.Tx, ids.ShortID, error) {
-	assetID, err := s.vm.lookupAssetID(args.AssetID)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	to, err := avax.ParseServiceAddress(s.vm, args.To)
-	if err != nil {
-		return nil, ids.ShortEmpty, fmt.Errorf("problem parsing to address %q: %w", args.To, err)
-	}
-
-	payloadBytes, err := formatting.Decode(args.Encoding, args.Payload)
-	if err != nil {
-		return nil, ids.ShortEmpty, fmt.Errorf("problem decoding payload bytes: %w", err)
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	feeUTXOs, feeKc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(feeKc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(feeKc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amountsSpent, ins, secpKeys, err := s.vm.Spend(
-		feeUTXOs,
-		feeKc,
-		map[ids.ID]uint64{
-			s.vm.feeAssetID: s.vm.TxFee,
-		},
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	outs := []*avax.TransferableOutput{}
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent > s.vm.TxFee {
-		outs = append(outs, &avax.TransferableOutput{
-			Asset: avax.Asset{ID: s.vm.feeAssetID},
-			Out: &secp256k1fx.TransferOutput{
-				Amt: amountSpent - s.vm.TxFee,
-				OutputOwners: secp256k1fx.OutputOwners{
-					Locktime:  0,
-					Threshold: 1,
-					Addrs:     []ids.ShortID{changeAddr},
-				},
-			},
-		})
-	}
-
-	// Get all UTXOs/keys
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, nil)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	ops, nftKeys, err := s.vm.MintNFT(
-		utxos,
-		kc,
-		assetID,
-		payloadBytes,
-		to,
-	)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	tx := &txs.Tx{Unsigned: &txs.OperationTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		Ops: ops,
-	}}
-
-	codec := s.vm.parser.Codec()
-	if err := tx.SignSECP256K1Fx(codec, secpKeys); err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-	return tx, changeAddr, tx.SignNFTFx(codec, nftKeys)
-}
-
-// ImportArgs are arguments for passing into Import requests
-type ImportArgs struct {
-	// User that controls To
-	api.UserPass
-
-	// Chain the funds are coming from
-	SourceChain string `json:"sourceChain"`
-
-	// Address receiving the imported AVAX
-	To string `json:"to"`
-}
-
-// Import imports an asset to this chain from the P/C-Chain.
-// The AVAX must have already been exported from the P/C-Chain.
-// Returns the ID of the newly created atomic transaction
-func (s *Service) Import(_ *http.Request, args *ImportArgs, reply *api.JSONTxID) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "import"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, err := s.buildImport(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	return nil
-}
-
-func (s *Service) buildImport(args *ImportArgs) (*txs.Tx, error) {
-	chainID, err := s.vm.ctx.BCLookup.Lookup(args.SourceChain)
-	if err != nil {
-		return nil, fmt.Errorf("problem parsing chainID %q: %w", args.SourceChain, err)
-	}
-
-	to, err := avax.ParseServiceAddress(s.vm, args.To)
-	if err != nil {
-		return nil, fmt.Errorf("problem parsing to address %q: %w", args.To, err)
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	atomicUTXOs, _, _, err := avax.GetAtomicUTXOs(
-		s.vm.ctx.SharedMemory,
-		s.vm.parser.Codec(),
-		chainID,
-		kc.Addrs,
-		ids.ShortEmpty,
-		ids.Empty,
-		int(maxPageSize),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("problem retrieving user's atomic UTXOs: %w", err)
-	}
-
-	amountsSpent, importInputs, importKeys, err := s.vm.SpendAll(atomicUTXOs, kc)
-	if err != nil {
-		return nil, err
-	}
-
-	ins := []*avax.TransferableInput{}
-	keys := [][]*secp256k1.PrivateKey{}
-
-	if amountSpent := amountsSpent[s.vm.feeAssetID]; amountSpent < s.vm.TxFee {
-		var localAmountsSpent map[ids.ID]uint64
-		localAmountsSpent, ins, keys, err = s.vm.Spend(
-			utxos,
-			kc,
-			map[ids.ID]uint64{
-				s.vm.feeAssetID: s.vm.TxFee - amountSpent,
-			},
-		)
-		if err != nil {
-			return nil, err
-		}
-		for asset, amount := range localAmountsSpent {
-			newAmount, err := safemath.Add(amountsSpent[asset], amount)
-			if err != nil {
-				return nil, fmt.Errorf("problem calculating required spend amount: %w", err)
-			}
-			amountsSpent[asset] = newAmount
-		}
-	}
-
-	// Because we ensured that we had enough inputs for the fee, we can
-	// safely just remove it without concern for underflow.
-	amountsSpent[s.vm.feeAssetID] -= s.vm.TxFee
-
-	keys = append(keys, importKeys...)
-
-	outs := []*avax.TransferableOutput{}
-	for assetID, amount := range amountsSpent {
-		if amount > 0 {
-			outs = append(outs, &avax.TransferableOutput{
-				Asset: avax.Asset{ID: assetID},
-				Out: &secp256k1fx.TransferOutput{
-					Amt: amount,
-					OutputOwners: secp256k1fx.OutputOwners{
-						Locktime:  0,
-						Threshold: 1,
-						Addrs:     []ids.ShortID{to},
-					},
-				},
-			})
-		}
-	}
-	avax.SortTransferableOutputs(outs, s.vm.parser.Codec())
-
-	tx := &txs.Tx{Unsigned: &txs.ImportTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		SourceChain: chainID,
-		ImportedIns: importInputs,
-	}}
-	return tx, tx.SignSECP256K1Fx(s.vm.parser.Codec(), keys)
-}
-
-// ExportArgs are arguments for passing into ExportAVA requests
-type ExportArgs struct {
-	// User, password, from addrs, change addr
-	api.JSONSpendHeader
-	// Amount of nAVAX to send
-	Amount avajson.Uint64 `json:"amount"`
-
-	// Chain the funds are going to. Optional. Used if To address does not include the chainID.
-	TargetChain string `json:"targetChain"`
-
-	// ID of the address that will receive the AVAX. This address may include the
-	// chainID, which is used to determine what the destination chain is.
-	To string `json:"to"`
-
-	AssetID string `json:"assetID"`
-}
-
-// Export sends an asset from this chain to the P/C-Chain.
-// After this tx is accepted, the AVAX must be imported to the P/C-chain with an importTx.
-// Returns the ID of the newly created atomic transaction
-func (s *Service) Export(_ *http.Request, args *ExportArgs, reply *api.JSONTxIDChangeAddr) error {
-	s.vm.ctx.Log.Warn("deprecated API called",
-		zap.String("service", "avm"),
-		zap.String("method", "export"),
-		logging.UserString("username", args.Username),
-	)
-
-	tx, changeAddr, err := s.buildExport(args)
-	if err != nil {
-		return err
-	}
-
-	txID, err := s.vm.issueTxFromRPC(tx)
-	if err != nil {
-		return fmt.Errorf("problem issuing transaction: %w", err)
-	}
-
-	reply.TxID = txID
-	reply.ChangeAddr, err = s.vm.FormatLocalAddress(changeAddr)
-	return err
-}
-
-func (s *Service) buildExport(args *ExportArgs) (*txs.Tx, ids.ShortID, error) {
-	// Parse the asset ID
-	assetID, err := s.vm.lookupAssetID(args.AssetID)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Get the chainID and parse the to address
-	chainID, to, err := s.vm.ParseAddress(args.To)
-	if err != nil {
-		chainID, err = s.vm.ctx.BCLookup.Lookup(args.TargetChain)
-		if err != nil {
-			return nil, ids.ShortEmpty, err
-		}
-		to, err = ids.ShortFromString(args.To)
-		if err != nil {
-			return nil, ids.ShortEmpty, err
-		}
-	}
-
-	if args.Amount == 0 {
-		return nil, ids.ShortEmpty, errZeroAmount
-	}
-
-	// Parse the from addresses
-	fromAddrs, err := avax.ParseServiceAddresses(s.vm, args.From)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	s.vm.ctx.Lock.Lock()
-	defer s.vm.ctx.Lock.Unlock()
-
-	// Get the UTXOs/keys for the from addresses
-	utxos, kc, err := s.vm.LoadUser(args.Username, args.Password, fromAddrs)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	// Parse the change address.
-	if len(kc.Keys) == 0 {
-		return nil, ids.ShortEmpty, errNoKeys
-	}
-	changeAddr, err := s.vm.selectChangeAddr(kc.Keys[0].PublicKey().Address(), args.ChangeAddr)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	amounts := map[ids.ID]uint64{}
-	if assetID == s.vm.feeAssetID {
-		amountWithFee, err := safemath.Add(uint64(args.Amount), s.vm.TxFee)
-		if err != nil {
-			return nil, ids.ShortEmpty, fmt.Errorf("problem calculating required spend amount: %w", err)
-		}
-		amounts[s.vm.feeAssetID] = amountWithFee
-	} else {
-		amounts[s.vm.feeAssetID] = s.vm.TxFee
-		amounts[assetID] = uint64(args.Amount)
-	}
-
-	amountsSpent, ins, keys, err := s.vm.Spend(utxos, kc, amounts)
-	if err != nil {
-		return nil, ids.ShortEmpty, err
-	}
-
-	exportOuts := []*avax.TransferableOutput{{
-		Asset: avax.Asset{ID: assetID},
-		Out: &secp256k1fx.TransferOutput{
-			Amt: uint64(args.Amount),
-			OutputOwners: secp256k1fx.OutputOwners{
-				Locktime:  0,
-				Threshold: 1,
-				Addrs:     []ids.ShortID{to},
-			},
-		},
-	}}
-
-	outs := []*avax.TransferableOutput{}
-	for assetID, amountSpent := range amountsSpent {
-		amountToSend := amounts[assetID]
-		if amountSpent > amountToSend {
-			outs = append(outs, &avax.TransferableOutput{
-				Asset: avax.Asset{ID: assetID},
-				Out: &secp256k1fx.TransferOutput{
-					Amt: amountSpent - amountToSend,
-					OutputOwners: secp256k1fx.OutputOwners{
-						Locktime:  0,
-						Threshold: 1,
-						Addrs:     []ids.ShortID{changeAddr},
-					},
-				},
-			})
-		}
-	}
-
-	codec := s.vm.parser.Codec()
-	avax.SortTransferableOutputs(outs, codec)
-
-	tx := &txs.Tx{Unsigned: &txs.ExportTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    s.vm.ctx.NetworkID,
-			BlockchainID: s.vm.ctx.ChainID,
-			Outs:         outs,
-			Ins:          ins,
-		}},
-		DestinationChain: chainID,
-		ExportedOuts:     exportOuts,
-	}}
-	return tx, changeAddr, tx.SignSECP256K1Fx(codec, keys)
 }
