+++ vms/avm/service_test.go
@@ -31,7 +31,6 @@ import (
 	"github.com/ava-labs/avalanchego/utils/units"
 	"github.com/ava-labs/avalanchego/vms/avm/block"
 	"github.com/ava-labs/avalanchego/vms/avm/block/executor/executormock"
-	"github.com/ava-labs/avalanchego/vms/avm/config"
 	"github.com/ava-labs/avalanchego/vms/avm/state/statemock"
 	"github.com/ava-labs/avalanchego/vms/avm/txs"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
@@ -2339,550 +2338,6 @@ func TestGetBalance(t *testing.T) {
 	}, &reply))
 
 	require.Equal(startBalance, uint64(reply.Balance))
-}
-
-func TestCreateFixedCapAsset(t *testing.T) {
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require := require.New(t)
-
-			env := setup(t, &envConfig{
-				fork:             upgradetest.Durango,
-				isCustomFeeAsset: !tc.avaxAsset,
-				keystoreUsers: []*user{{
-					username:    username,
-					password:    password,
-					initialKeys: keys,
-				}},
-			})
-			service := &Service{vm: env.vm}
-			env.vm.ctx.Lock.Unlock()
-
-			reply := AssetIDChangeAddr{}
-			addrStr, err := env.vm.FormatLocalAddress(keys[0].PublicKey().Address())
-			require.NoError(err)
-
-			changeAddrStr, err := env.vm.FormatLocalAddress(testChangeAddr)
-			require.NoError(err)
-			_, fromAddrsStr := sampleAddrs(t, env.vm.AddressManager, addrs)
-
-			require.NoError(service.CreateFixedCapAsset(nil, &CreateAssetArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{From: fromAddrsStr},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-				},
-				Name:         "testAsset",
-				Symbol:       "TEST",
-				Denomination: 1,
-				InitialHolders: []*Holder{{
-					Amount:  123456789,
-					Address: addrStr,
-				}},
-			}, &reply))
-			require.Equal(changeAddrStr, reply.ChangeAddr)
-		})
-	}
-}
-
-func TestCreateVariableCapAsset(t *testing.T) {
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require := require.New(t)
-
-			env := setup(t, &envConfig{
-				fork:             upgradetest.Durango,
-				isCustomFeeAsset: !tc.avaxAsset,
-				keystoreUsers: []*user{{
-					username:    username,
-					password:    password,
-					initialKeys: keys,
-				}},
-			})
-			service := &Service{vm: env.vm}
-			env.vm.ctx.Lock.Unlock()
-
-			reply := AssetIDChangeAddr{}
-			minterAddrStr, err := env.vm.FormatLocalAddress(keys[0].PublicKey().Address())
-			require.NoError(err)
-			_, fromAddrsStr := sampleAddrs(t, env.vm.AddressManager, addrs)
-			changeAddrStr := fromAddrsStr[0]
-
-			require.NoError(service.CreateVariableCapAsset(nil, &CreateAssetArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{From: fromAddrsStr},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-				},
-				Name:   "test asset",
-				Symbol: "TEST",
-				MinterSets: []Owners{
-					{
-						Threshold: 1,
-						Minters: []string{
-							minterAddrStr,
-						},
-					},
-				},
-			}, &reply))
-			require.Equal(changeAddrStr, reply.ChangeAddr)
-
-			buildAndAccept(require, env.vm, env.issuer, reply.AssetID)
-
-			createdAssetID := reply.AssetID.String()
-			// Test minting of the created variable cap asset
-			mintArgs := &MintArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-				},
-				Amount:  200,
-				AssetID: createdAssetID,
-				To:      minterAddrStr, // Send newly minted tokens to this address
-			}
-			mintReply := &api.JSONTxIDChangeAddr{}
-			require.NoError(service.Mint(nil, mintArgs, mintReply))
-			require.Equal(changeAddrStr, mintReply.ChangeAddr)
-
-			buildAndAccept(require, env.vm, env.issuer, mintReply.TxID)
-
-			sendArgs := &SendArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{From: []string{minterAddrStr}},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-				},
-				SendOutput: SendOutput{
-					Amount:  200,
-					AssetID: createdAssetID,
-					To:      fromAddrsStr[0],
-				},
-			}
-			sendReply := &api.JSONTxIDChangeAddr{}
-			require.NoError(service.Send(nil, sendArgs, sendReply))
-			require.Equal(changeAddrStr, sendReply.ChangeAddr)
-		})
-	}
-}
-
-func TestNFTWorkflow(t *testing.T) {
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require := require.New(t)
-
-			env := setup(t, &envConfig{
-				fork:             upgradetest.Durango,
-				isCustomFeeAsset: !tc.avaxAsset,
-				keystoreUsers: []*user{{
-					username:    username,
-					password:    password,
-					initialKeys: keys,
-				}},
-			})
-			service := &Service{vm: env.vm}
-			env.vm.ctx.Lock.Unlock()
-
-			fromAddrs, fromAddrsStr := sampleAddrs(t, env.vm.AddressManager, addrs)
-
-			// Test minting of the created variable cap asset
-			addrStr, err := env.vm.FormatLocalAddress(keys[0].PublicKey().Address())
-			require.NoError(err)
-
-			createArgs := &CreateNFTAssetArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{From: fromAddrsStr},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: fromAddrsStr[0]},
-				},
-				Name:   "BIG COIN",
-				Symbol: "COIN",
-				MinterSets: []Owners{
-					{
-						Threshold: 1,
-						Minters: []string{
-							addrStr,
-						},
-					},
-				},
-			}
-			createReply := &AssetIDChangeAddr{}
-			require.NoError(service.CreateNFTAsset(nil, createArgs, createReply))
-			require.Equal(fromAddrsStr[0], createReply.ChangeAddr)
-
-			buildAndAccept(require, env.vm, env.issuer, createReply.AssetID)
-
-			// Key: Address
-			// Value: AVAX balance
-			balances := map[ids.ShortID]uint64{}
-			for _, addr := range addrs { // get balances for all addresses
-				addrStr, err := env.vm.FormatLocalAddress(addr)
-				require.NoError(err)
-
-				reply := &GetBalanceReply{}
-				require.NoError(service.GetBalance(nil,
-					&GetBalanceArgs{
-						Address: addrStr,
-						AssetID: env.vm.feeAssetID.String(),
-					},
-					reply,
-				))
-
-				balances[addr] = uint64(reply.Balance)
-			}
-
-			fromAddrsTotalBalance := uint64(0)
-			for _, addr := range fromAddrs {
-				fromAddrsTotalBalance += balances[addr]
-			}
-
-			fromAddrsStartBalance := startBalance * uint64(len(fromAddrs))
-			require.Equal(fromAddrsStartBalance-env.vm.TxFee, fromAddrsTotalBalance)
-
-			assetID := createReply.AssetID
-			payload, err := formatting.Encode(formatting.Hex, []byte{1, 2, 3, 4, 5})
-			require.NoError(err)
-			mintArgs := &MintNFTArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: fromAddrsStr[0]},
-				},
-				AssetID:  assetID.String(),
-				Payload:  payload,
-				To:       addrStr,
-				Encoding: formatting.Hex,
-			}
-			mintReply := &api.JSONTxIDChangeAddr{}
-
-			require.NoError(service.MintNFT(nil, mintArgs, mintReply))
-			require.Equal(fromAddrsStr[0], createReply.ChangeAddr)
-
-			// Accept the transaction so that we can send the newly minted NFT
-			buildAndAccept(require, env.vm, env.issuer, mintReply.TxID)
-
-			sendArgs := &SendNFTArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: fromAddrsStr[0]},
-				},
-				AssetID: assetID.String(),
-				GroupID: 0,
-				To:      addrStr,
-			}
-			sendReply := &api.JSONTxIDChangeAddr{}
-			require.NoError(service.SendNFT(nil, sendArgs, sendReply))
-			require.Equal(fromAddrsStr[0], sendReply.ChangeAddr)
-		})
-	}
-}
-
-func TestImportExportKey(t *testing.T) {
-	require := require.New(t)
-
-	env := setup(t, &envConfig{
-		fork: upgradetest.Durango,
-		keystoreUsers: []*user{{
-			username: username,
-			password: password,
-		}},
-	})
-	service := &Service{vm: env.vm}
-	env.vm.ctx.Lock.Unlock()
-
-	sk, err := secp256k1.NewPrivateKey()
-	require.NoError(err)
-
-	importArgs := &ImportKeyArgs{
-		UserPass: api.UserPass{
-			Username: username,
-			Password: password,
-		},
-		PrivateKey: sk,
-	}
-	importReply := &api.JSONAddress{}
-	require.NoError(service.ImportKey(nil, importArgs, importReply))
-
-	addrStr, err := env.vm.FormatLocalAddress(sk.PublicKey().Address())
-	require.NoError(err)
-	exportArgs := &ExportKeyArgs{
-		UserPass: api.UserPass{
-			Username: username,
-			Password: password,
-		},
-		Address: addrStr,
-	}
-	exportReply := &ExportKeyReply{}
-	require.NoError(service.ExportKey(nil, exportArgs, exportReply))
-	require.Equal(sk.Bytes(), exportReply.PrivateKey.Bytes())
-}
-
-func TestImportAVMKeyNoDuplicates(t *testing.T) {
-	require := require.New(t)
-
-	env := setup(t, &envConfig{
-		fork: upgradetest.Durango,
-		keystoreUsers: []*user{{
-			username: username,
-			password: password,
-		}},
-	})
-	service := &Service{vm: env.vm}
-	env.vm.ctx.Lock.Unlock()
-
-	sk, err := secp256k1.NewPrivateKey()
-	require.NoError(err)
-	args := ImportKeyArgs{
-		UserPass: api.UserPass{
-			Username: username,
-			Password: password,
-		},
-		PrivateKey: sk,
-	}
-	reply := api.JSONAddress{}
-	require.NoError(service.ImportKey(nil, &args, &reply))
-
-	expectedAddress, err := env.vm.FormatLocalAddress(sk.PublicKey().Address())
-	require.NoError(err)
-
-	require.Equal(expectedAddress, reply.Address)
-
-	reply2 := api.JSONAddress{}
-	require.NoError(service.ImportKey(nil, &args, &reply2))
-
-	require.Equal(expectedAddress, reply2.Address)
-
-	addrsArgs := api.UserPass{
-		Username: username,
-		Password: password,
-	}
-	addrsReply := api.JSONAddresses{}
-	require.NoError(service.ListAddresses(nil, &addrsArgs, &addrsReply))
-
-	require.Len(addrsReply.Addresses, 1)
-	require.Equal(expectedAddress, addrsReply.Addresses[0])
-}
-
-func TestSend(t *testing.T) {
-	require := require.New(t)
-
-	env := setup(t, &envConfig{
-		fork: upgradetest.Durango,
-		keystoreUsers: []*user{{
-			username:    username,
-			password:    password,
-			initialKeys: keys,
-		}},
-	})
-	service := &Service{vm: env.vm}
-	env.vm.ctx.Lock.Unlock()
-
-	assetID := env.genesisTx.ID()
-	addr := keys[0].PublicKey().Address()
-
-	addrStr, err := env.vm.FormatLocalAddress(addr)
-	require.NoError(err)
-	changeAddrStr, err := env.vm.FormatLocalAddress(testChangeAddr)
-	require.NoError(err)
-	_, fromAddrsStr := sampleAddrs(t, env.vm.AddressManager, addrs)
-
-	args := &SendArgs{
-		JSONSpendHeader: api.JSONSpendHeader{
-			UserPass: api.UserPass{
-				Username: username,
-				Password: password,
-			},
-			JSONFromAddrs:  api.JSONFromAddrs{From: fromAddrsStr},
-			JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-		},
-		SendOutput: SendOutput{
-			Amount:  500,
-			AssetID: assetID.String(),
-			To:      addrStr,
-		},
-	}
-	reply := &api.JSONTxIDChangeAddr{}
-	require.NoError(service.Send(nil, args, reply))
-	require.Equal(changeAddrStr, reply.ChangeAddr)
-
-	buildAndAccept(require, env.vm, env.issuer, reply.TxID)
-}
-
-func TestSendMultiple(t *testing.T) {
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require := require.New(t)
-
-			env := setup(t, &envConfig{
-				isCustomFeeAsset: !tc.avaxAsset,
-				keystoreUsers: []*user{{
-					username:    username,
-					password:    password,
-					initialKeys: keys,
-				}},
-				vmStaticConfig: &config.Config{
-					Upgrades: upgradetest.GetConfig(upgradetest.Durango),
-				},
-			})
-			service := &Service{vm: env.vm}
-			env.vm.ctx.Lock.Unlock()
-
-			assetID := env.genesisTx.ID()
-			addr := keys[0].PublicKey().Address()
-
-			addrStr, err := env.vm.FormatLocalAddress(addr)
-			require.NoError(err)
-			changeAddrStr, err := env.vm.FormatLocalAddress(testChangeAddr)
-			require.NoError(err)
-			_, fromAddrsStr := sampleAddrs(t, env.vm.AddressManager, addrs)
-
-			args := &SendMultipleArgs{
-				JSONSpendHeader: api.JSONSpendHeader{
-					UserPass: api.UserPass{
-						Username: username,
-						Password: password,
-					},
-					JSONFromAddrs:  api.JSONFromAddrs{From: fromAddrsStr},
-					JSONChangeAddr: api.JSONChangeAddr{ChangeAddr: changeAddrStr},
-				},
-				Outputs: []SendOutput{
-					{
-						Amount:  500,
-						AssetID: assetID.String(),
-						To:      addrStr,
-					},
-					{
-						Amount:  1000,
-						AssetID: assetID.String(),
-						To:      addrStr,
-					},
-				},
-			}
-			reply := &api.JSONTxIDChangeAddr{}
-			require.NoError(service.SendMultiple(nil, args, reply))
-			require.Equal(changeAddrStr, reply.ChangeAddr)
-
-			buildAndAccept(require, env.vm, env.issuer, reply.TxID)
-		})
-	}
-}
-
-func TestCreateAndListAddresses(t *testing.T) {
-	require := require.New(t)
-
-	env := setup(t, &envConfig{
-		fork: upgradetest.Durango,
-		keystoreUsers: []*user{{
-			username: username,
-			password: password,
-		}},
-	})
-	service := &Service{vm: env.vm}
-	env.vm.ctx.Lock.Unlock()
-
-	createArgs := &api.UserPass{
-		Username: username,
-		Password: password,
-	}
-	createReply := &api.JSONAddress{}
-
-	require.NoError(service.CreateAddress(nil, createArgs, createReply))
-
-	newAddr := createReply.Address
-
-	listArgs := &api.UserPass{
-		Username: username,
-		Password: password,
-	}
-	listReply := &api.JSONAddresses{}
-
-	require.NoError(service.ListAddresses(nil, listArgs, listReply))
-	require.Contains(listReply.Addresses, newAddr)
-}
-
-func TestImport(t *testing.T) {
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require := require.New(t)
-
-			env := setup(t, &envConfig{
-				fork:             upgradetest.Durango,
-				isCustomFeeAsset: !tc.avaxAsset,
-				keystoreUsers: []*user{{
-					username:    username,
-					password:    password,
-					initialKeys: keys,
-				}},
-			})
-			service := &Service{vm: env.vm}
-			env.vm.ctx.Lock.Unlock()
-
-			assetID := env.genesisTx.ID()
-			addr0 := keys[0].PublicKey().Address()
-
-			utxo := &avax.UTXO{
-				UTXOID: avax.UTXOID{TxID: ids.Empty},
-				Asset:  avax.Asset{ID: assetID},
-				Out: &secp256k1fx.TransferOutput{
-					Amt: 7,
-					OutputOwners: secp256k1fx.OutputOwners{
-						Threshold: 1,
-						Addrs:     []ids.ShortID{addr0},
-					},
-				},
-			}
-			utxoBytes, err := env.vm.parser.Codec().Marshal(txs.CodecVersion, utxo)
-			require.NoError(err)
-
-			peerSharedMemory := env.sharedMemory.NewSharedMemory(constants.PlatformChainID)
-			utxoID := utxo.InputID()
-			require.NoError(peerSharedMemory.Apply(map[ids.ID]*atomic.Requests{
-				env.vm.ctx.ChainID: {
-					PutRequests: []*atomic.Element{{
-						Key:   utxoID[:],
-						Value: utxoBytes,
-						Traits: [][]byte{
-							addr0.Bytes(),
-						},
-					}},
-				},
-			}))
-
-			addrStr, err := env.vm.FormatLocalAddress(keys[0].PublicKey().Address())
-			require.NoError(err)
-			args := &ImportArgs{
-				UserPass: api.UserPass{
-					Username: username,
-					Password: password,
-				},
-				SourceChain: "P",
-				To:          addrStr,
-			}
-			reply := &api.JSONTxID{}
-			require.NoError(service.Import(nil, args, reply))
-		})
-	}
 }
 
 func TestServiceGetBlock(t *testing.T) {
