+++ vms/avm/vm.go
@@ -38,8 +38,6 @@ import (
 	"github.com/ava-labs/avalanchego/vms/avm/utxo"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
 	"github.com/ava-labs/avalanchego/vms/components/index"
-	"github.com/ava-labs/avalanchego/vms/components/keystore"
-	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 	"github.com/ava-labs/avalanchego/vms/txs/mempool"
 
 	blockbuilder "github.com/ava-labs/avalanchego/vms/avm/block/builder"
@@ -575,54 +573,6 @@ func (vm *VM) initState(tx *txs.Tx) {
 	for _, utxo := range tx.UTXOs() {
 		vm.state.AddUTXO(utxo)
 	}
-}
-
-// LoadUser returns:
-// 1) The UTXOs that reference one or more addresses controlled by the given user
-// 2) A keychain that contains this user's keys
-// If [addrsToUse] has positive length, returns UTXOs that reference one or more
-// addresses controlled by the given user that are also in [addrsToUse].
-func (vm *VM) LoadUser(
-	username string,
-	password string,
-	addrsToUse set.Set[ids.ShortID],
-) (
-	[]*avax.UTXO,
-	*secp256k1fx.Keychain,
-	error,
-) {
-	user, err := keystore.NewUserFromKeystore(vm.ctx.Keystore, username, password)
-	if err != nil {
-		return nil, nil, err
-	}
-	// Drop any potential error closing the database to report the original
-	// error
-	defer user.Close()
-
-	kc, err := keystore.GetKeychain(user, addrsToUse)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	utxos, err := avax.GetAllUTXOs(vm.state, kc.Addresses())
-	if err != nil {
-		return nil, nil, fmt.Errorf("problem retrieving user's UTXOs: %w", err)
-	}
-
-	return utxos, kc, user.Close()
-}
-
-// selectChangeAddr returns the change address to be used for [kc] when [changeAddr] is given
-// as the optional change address argument
-func (vm *VM) selectChangeAddr(defaultAddr ids.ShortID, changeAddr string) (ids.ShortID, error) {
-	if changeAddr == "" {
-		return defaultAddr, nil
-	}
-	addr, err := avax.ParseServiceAddress(vm, changeAddr)
-	if err != nil {
-		return ids.ShortID{}, fmt.Errorf("couldn't parse changeAddr: %w", err)
-	}
-	return addr, nil
 }
 
 // lookupAssetID looks for an ID aliased by [asset] and if it fails
