+++ vms/avm/vm_benchmark_test.go
@@ -4,7 +4,6 @@
 package avm
 
 import (
-	"fmt"
 	"math/rand"
 	"testing"
 
@@ -14,52 +13,8 @@ import (
 	"github.com/ava-labs/avalanchego/upgrade/upgradetest"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
-	"github.com/ava-labs/avalanchego/vms/components/keystore"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 )
-
-func BenchmarkLoadUser(b *testing.B) {
-	runLoadUserBenchmark := func(b *testing.B, numKeys int) {
-		require := require.New(b)
-
-		env := setup(b, &envConfig{
-			fork: upgradetest.Latest,
-			keystoreUsers: []*user{{
-				username: username,
-				password: password,
-			}},
-		})
-		defer env.vm.ctx.Lock.Unlock()
-
-		user, err := keystore.NewUserFromKeystore(env.vm.ctx.Keystore, username, password)
-		require.NoError(err)
-
-		keys, err := keystore.NewKeys(user, numKeys)
-		require.NoError(err)
-
-		b.ResetTimer()
-
-		fromAddrs := set.Set[ids.ShortID]{}
-		for n := 0; n < b.N; n++ {
-			addrIndex := n % numKeys
-			fromAddrs.Clear()
-			fromAddrs.Add(keys[addrIndex].PublicKey().Address())
-			_, _, err := env.vm.LoadUser(username, password, fromAddrs)
-			require.NoError(err)
-		}
-
-		b.StopTimer()
-
-		require.NoError(user.Close())
-	}
-
-	benchmarkSize := []int{10, 100, 1000, 5000}
-	for _, numKeys := range benchmarkSize {
-		b.Run(fmt.Sprintf("NumKeys=%d", numKeys), func(b *testing.B) {
-			runLoadUserBenchmark(b, numKeys)
-		})
-	}
-}
 
 // getAllUTXOsBenchmark is a helper func to benchmark the GetAllUTXOs depending on the size
 func getAllUTXOsBenchmark(b *testing.B, utxoCount int, randSrc rand.Source) {
