--- vms/components/keystore/codec.go
@@ -1,34 +1,0 @@
-// Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package keystore
-
-import (
-	"errors"
-	"math"
-
-	"github.com/ava-labs/avalanchego/codec"
-	"github.com/ava-labs/avalanchego/codec/linearcodec"
-)
-
-const CodecVersion = 0
-
-var (
-	Codec       codec.Manager
-	LegacyCodec codec.Manager
-)
-
-func init() {
-	c := linearcodec.NewDefault()
-	Codec = codec.NewDefaultManager()
-	lc := linearcodec.NewDefault()
-	LegacyCodec = codec.NewManager(math.MaxInt32)
-
-	err := errors.Join(
-		Codec.RegisterCodec(CodecVersion, c),
-		LegacyCodec.RegisterCodec(CodecVersion, lc),
-	)
-	if err != nil {
-		panic(err)
-	}
-}
